move `ClockIcon` to its own file
